requirements.txt
future
requests
django
djangorestframework
jsonfield
